+++ src/components/PlayerAchievements.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './PlayerAchievements.scss'
 
 const PlayerAchivements = ({ achievementsList = null }) => {
-
+  
   const formatDateString = (dateString) => {
     const date = new Date(dateString);
     const formatter = new Intl.DateTimeFormat('en-US', {
@@ -20,8 +20,10 @@ const PlayerAchivements = ({ achievementsList = null }) => {
         <ul>
           {achievementsList.map((achievement,idx) => (
             <li key={idx}>
-              <div className="icon">
-                <i className={'fa-solid fa-medal tier-'+achievement.achievement_tier.toLowerCase()}></i>
+              <div className={`badge`}>
+                <div className={`icon tier-` + achievement.achievement_tier.toLowerCase()}>
+                  <img src={`/assets/badges/${achievement.sysname}.png`} alt={achievement.achievement_name} />
+                </div>
               </div>
               <div className="details">
                 <h4 className='achievement-name'>{achievement.achievement_name} <small>(Tier {achievement.achievement_tier})</small></h4>